requirements/testing.txt
pytest==3.6.0
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-mock==1.10.0
pytest-sugar==0.9.1
